fix: arm64 build and old kernel versions

+++ driver/modern_bpf/programs/attached/events/sched_process_fork.bpf.c
@@ -233,7 +233,7 @@ int BPF_PROG(t2_sched_p_fork,
 	}
 
 	/* Parameter 21: pid_namespace init task start_time monotonic time in ns (type: PT_UINT64) */
-	auxmap__store_u64_param(auxmap, (u64)extract__task_pidns_start_time(child, PIDTYPE_TGID, 0));
+	auxmap__store_u64_param(auxmap, extract__task_pidns_start_time(child, PIDTYPE_TGID, 0));
 
 	/*=============================== COLLECT PARAMETERS  ===========================*/
 
@@ -241,7 +241,5 @@ int BPF_PROG(t2_sched_p_fork,
 
 	auxmap__submit_event(auxmap);
 	return 0;
-
 }
-
 #endif /* CAPTURE_SCHED_PROC_EXEC */
